- add dnssec-tools-config.h - fix compiler warning: change length param to size_t
git-svn-id: https://www.dnssec-tools.org/svn/dnssec-tools/trunk@1575 c4029f9c-b713-0410-a448-fae9e31a314a

+++ dnssec-tools/lib/libval/val_verify.c
@@ -6,6 +6,7 @@
  *
  * This is the implementation file for the verifier.
  */
+#include "../../dnssec-tools-config.h"
 
 #include <stdio.h>
 #include <stdlib.h>
@@ -250,7 +251,7 @@ static int predict_sigbuflength (  struct rrset_rec *rr_set,
 }
 
 static int make_sigfield (  u_int8_t            **field,
-                        int                 *field_length,
+                        size_t              *field_length,
                         struct rrset_rec    *rr_set,
                         struct rr_rec       *rr_sig,
                         int                 is_a_wildcard)
